requirements.txt
requests
ffmpeg-python
pydub
google-cloud-speech
google-cloud-texttospeech
pvporcupine
openai
mypy
